requirements.txt
tensorflow-gpu
keras
numpy
pandas
tqdm
sklearn
torchmetrics
torch
torchvision
matplotlib
opencv-python
albumentations
pytorch_lightning
collections
torchviz